fin: fix everything that was wrong with the arduino code

+++ src/arduino/src/Handshake/Handshake.cpp
@@ -1,6 +1,6 @@
 #include "Handshake.h"
 
-Handshake::Handshake(int pin)
+Handshake::Handshake(byte pin)
 {
   this->_pin = pin;
   init();
